Setup boilerplate for login and sessions
This is just a starting point really, there is lots of room for improvement still, for example using redis for session later, etc.

The secret key isn't exactly kept secret for now so it's just a default setting stored in GIT in the .ini files. The CMS is nowhere near production ready anyway, so this is OK for the time being, but later we'll need a way to be able set the secret key, but not store it on GIT.