Fix: import * from models didn't work
Python expects the list of models in __all__ to be a list of strings and not objects, or you get an error if you try to do "from pyramidcms.db.models import *" in pcms shell.

When we need to get to the actual model class, we can use getattr() on the module to get a reference to it.

When inside the module itself, use getattr on sys.modules[__name__], which is the current module.

While updating the dumpdata command, made it slightly more readable, by not having the nested list comprehensions.

Closes #13