Don't make flake8 a separate tox environment
This makes it a bit harder to deal with Jenkins, as each environment in a multi-configuration project will then have a different set of reports.

For example the py34 step doesn't have a flake8 report, and the flake8 step doesn't have a junit or coverage report.

However, if we always run flake8 for each interpreter we support, though we are running flake8 multiple times, each jenkins sub project in a multi-configuration project will then have the same set of reports.