Run flake8 only once now
Decided to only run flake8 once and output to a file.

If the developer wants to see flake8 output, they can always just run the "flake8" command manually, so there isn't really a need to show flake8 output during tox.

The main purpose of a tox run is to generate reports for Jenkins anyway.