Try adding a .travis.yml file
We shouldn't have to specify the Python version as tox should create a virtualenv for the specified python versions in the tox.ini file anyway.

In this case, the python version in .travis.yml is just the python version we use to run tox on, not the tests which is defined in the tox.ini file.