Move command_help outside the main function
This is something I must have missed during code review, sometimes an inner function makes sense, but not in this case.

Having it as a regular function makes the code a bit easier to read, and test.

Also renamed the function to print_command_help and do some refactoring to make it a bit clearer.

I still don't like the long list comprehension in print_command_help, so left the TODO note in for now.