Refactor command list help code again
Decided it's best to have a function that returns the list of available commands instead (get_command_list), it reads nicer and it would be easier for testing too, and the amount of code moved to the main() method is minimal.

Then we can just call parser.print_help() to print the default help text directly rather than capturing it in a variable using parser.format_help(), then finally print the list of commands in the main() method instead of it's own method because it's only 2 lines of code.