Remove __init__.py from tests directory
The problem with having an __init__.py here is that the find_packages() function in setup.py detects this as a module and will include it.

This becomes a problem when you debian-package pyramidcms, it will end up installing two modules in the site-packages folder, "pyramidcms" and "tests".

If multiple packages had a tests module in it, this would clash and both Debian packages could not be installed at the same time.

Luckily nosetests doesn't seem to have a problem with removing __init__.py from the tests directory.