Authorization class *_list methods should return a filtered list, not True/False
This is similar to what TastyPie does for list authorization:

http://django-tastypie.readthedocs.org/en/latest/authorization.html#return-values-the-list-case

In the case of ACLAuthorization we either return all the items if we have the required permission, otherwise return an empty list.

There is the potential for an Authorization class to filter based on permission if it is designed this way however, which is more flexible.