Add setUp and tearDown to API base tests
What was happening, is tests that were modifying properties on the API metaclass were doing it in such a way it was affecting tests that run after it.

By backing up these API metaclass properties in the setUp() method before each test, they can be automatically restored in tearDown() and therefore not affecting other tests anymore.