Started browsable API renderer
In order to achieve different renderers (the browsable API renderer
for now, but with the possibility of maybe a json-api renderer later),
some refactoring of the API was needed.

Mostly this involved returning Bundle objects from API view methods,
rather than returning a dict.  This also meant a __json__ method had
to be added to the Bundle object, which gets used by the json renderer,
so that the json renderer still returns the original data as before.

Some rework was also done in collection_get, where it now creates
a Bundle for the entire set of items, which lines up a bit more to
what TastyPie does.  This was originally done in the paginator @reify
property, this property had to be removed as part of this refactor
and moved into the collection_get method itself.